1. change the file names in the unit tests as well.
