+++ unit_test/mocks/mock_freertos.cpp
@@ -11,7 +11,7 @@
  ************************************************************************************************************/
 
 //************************************************** Includes ************************************************
-#include "mock_FreeRTOS.h"
+#include "mock_freertos.h"
 
 //********************************************** Mock Definitions ********************************************
 QueueHandle_t xQueueCreateMutex( const uint8_t ucQueueType )
